Updates for timing and end game conditions
The time between questions is only 10 seconds instead of 30. this makes it easier to test. checkForEndGame(); has been moved to the timer function so it is only called once for every new question. The same crash happens when 3 people die. I think it might have to do with scene build settings